refactor: moving widget into their own package
