+++ core/src/main/java/dev/kingdomino/screen/NextDominoRenderManager.java
@@ -13,6 +13,8 @@ import dev.kingdomino.game.Domino;
 import dev.kingdomino.game.DraftInputHandler;
 import dev.kingdomino.game.GameManager;
 import dev.kingdomino.game.GameManager.GameState;
+import dev.kingdomino.screen.widgets.DominoActor;
+import dev.kingdomino.screen.widgets.WavyLabel;
 import dev.kingdomino.game.King;
 import dev.kingdomino.game.Turn;
