+++ core/src/main/java/dev/kingdomino/screen/widgets/WavyLabel.java
@@ -1,4 +1,4 @@
-package dev.kingdomino.screen;
+package dev.kingdomino.screen.widgets;
 
 import java.util.Random;
